+++ CHANGELOG.md
@@ -5,6 +5,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [1.1.0] - 2019-07-23
+### Changed
+- base model attributes switched underscore to camelcase
+### Added
+- create / update / destroy
+- response errors
+- .first method = for return first record :)
+### Fixed
+- Content-Type header in post
+### Removed
+- class-instance variable @response
 ## [1.0.3] - 2019-07-18
 ### Added
 - store @response in variable